@@ -46,7 +46,7 @@
     "tslint": "^3.5.0",
     "tslint-stylish": "^2.1.0-beta",
     "typedoc": "^0.5.1",
-    "typescript": "^2.0.0"
+    "typescript": "^2.1.0"
   },
   "dependencies": {
     "@reactivex/rxjs": "^5.0.0-rc.5",
